api-agents/shared/requirements-base.txt
# Base dependencies for all API agents
# Shared by every api-agent-* image via multi-stage Docker build

firebase-admin>=6.5,<7
google-cloud-firestore>=2.16,<3
google-cloud-kms>=2.21,<3
google-cloud-storage>=2.14,<3
cryptography>=42,<44
pyyaml>=6.0,<7
requests>=2.31,<3
httpx>=0.27,<1
deepagents>=0.4,<1
